Unit-test ModuleSpec, not models.ModuleVersion
